[APM] add filters support to apm latency, throughput, and error rate chart apis (#181359)
## Summary

Rational:
We'd like to embed APM visualizations across the observability solution,
particularly within the SLO alert details page at this time. SLO
configuration supports unified search filters. In order to ensure that
the data accurately reflects the SLO configuration, API dependencies for
APM visualizations must support filters.

This PR adds filters support to:

1. `GET
/internal/apm/services/{serviceName}/transactions/charts/latency`
2. `GET /internal/apm/services/{serviceName}/throughput`
3. `GET
/internal/apm/services/{serviceName}/transactions/charts/error_rate`

It is expected that consumers of the filters param send a serialized
object containing a `filter` or `must_not` clause to include on the
respective ES queries.

Internally, it is expected that these objects are created using the
`buildQueryFromFilters` helper exposed by `kbn/es-query`, passing the
`Filter` object from the unified search `SearchBar` as the the
parameter.

### Testing
This feature is not yet available in the UI
To test, I've added api integration tests for each api, as well as jest
tests for any helpers introduced.