fix(slo): remove assertion on deleted rollup documents (#181725)
Resolves https://github.com/elastic/kibana/issues/180982

## Summary

When deleting an SLO, we also start a delete_by_query on the rollup
documents. This request is done asynchronously in the background. The
serverless integration test asserts on the deletion of the rollup
documents but fails some time.
As the deletion of the rollup document causes no harms if not done, I'm
removing this assertion from the test.