Fixes Failing test: X-Pack Alerting API Integration Tests - Alerting - group4.x-pack/test/alerting_api_integration/spaces_only/tests/alerting/group4/alerts_as_data/install_resources·ts - Alerting alerts_as_data install alerts as data resources should install context specific alerts as data resources on startup (#181759)
Resolves https://github.com/elastic/kibana/issues/181296

## Summary

Inspecting the logs of the flaky runs, everything looks like it's
getting installed correctly but the test is checking for the existence
of the concrete index before it has a chance to get installed. Added a
retry in order to allow for more time to let the index get installed.

Flaky test runner:
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/5766

(cherry picked from commit 5b2300bf0e4580c0e777474cc1d281bc9a747639)

# Conflicts:
#	x-pack/test/alerting_api_integration/spaces_only/tests/alerting/group4/alerts_as_data/install_resources.ts