[data views] Runtime field creation - display index values instead of source values in preview pane  (#181246)
## Summary

The preview pane has two different groups - values relevant to the
currently edited field and other values in the same document.

For other values in the same doc, we were showing source values which
are often but not always the same as indexed values. This has been fixed
- its showing indexed values now.

As for the current field values, those were picking from _source or
indexed values but the code provided one as the fallback for another.
While this worked, the code didn't read as though it was doing the right
thing due to context not present in the code. Now, a runtime field
without a script obviously only returns a source value and the preview
of a mapped field only returns a mapped value.

```
PUT test
{
  "mappings": {
    "dynamic": false,
    "properties": {
      "percentage1": {
        "type": "long"
      },
      "percentage2": {
        "type": "long"
      },
      "percentage3": {
        "type": "long"
      }
    }
  }
}

PUT test/_doc/doc1
{
  "key": "value",
  "percentage1": 0.28,
  "percentage2": 0,
  "percentage3": 0.04,
  "percentage4": 0.21
}
```

Useful starting point for testing runtime field results -

```
GET test/_search
{
  "runtime_mappings": {
    "key.keyword": {
      "type": "keyword"
    }
  },
  "fields" : ["*"]
}
```

Closes https://github.com/elastic/kibana/issues/180742

## Release notes

The runtime field creation modal now shows indexed values instead of
source values in the preview pane.

---------

Co-authored-by: Julia Rechkunova <julia.rechkunova@gmail.com>
Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>