Try making risk engine painless scripts static. also remove category based weighting
the category based weighting, like all weighting, isn't used and we probably won't re-implement it in the next version of risk scoring

this commit probably doesn't work. i can't get integration tests running locally

I put the painless scripts in a directory and I load them from the file system at runtime.

We should figure out how to bundle the scripts possibly. I'm not sure that it matters a lot, but I would like any static analysis processes to be able to work with this code as usual. e.g. i'd like to get a static-time error if a file is missing vs needing to run the risk engine task to find out a file is missing