Dashboard filter by creator (#180147)
Part of https://github.com/elastic/kibana-team/issues/769

## Summary

This PR adds a "created by" filter to the dashboard listing page. We've
added the created_by field recently
[here](https://github.com/elastic/kibana/pull/179344) so only newly
created by interactive users dashboards will be populated with
`created_by`

### Demo 

<img width="1510" alt="Screenshot 2024-04-25 at 16 32 03"
src="https://github.com/elastic/kibana/assets/7784120/ac70ffea-05be-4d70-bbfa-c1cc9d5dd747">

<img width="1510" alt="Screenshot 2024-04-25 at 16 32 11"
src="https://github.com/elastic/kibana/assets/7784120/9889ccae-08d0-4fb6-977e-c42e85337e0e">


### Implementation notes

#### Suggesting users

We suggest only the users that have at least one dashboard. We do this
by aggregating dashboard saved objects by `created_by`. Because we don't
have server-side pagination or sorting yet, in this initial version, I
decided to do this client-side to keep it as simple as possible for now.

Initially, I did this server-side inside a dashboard plugin as an
internal API route. In the long term, this should probably be moved to
content management in some form, but it is not clear how it should look
yet. This is the commit where I reverted and moved it to client-side
https://github.com/elastic/kibana/pull/180147/commits/06316ee8f2fa9a97dd9bf29ae090fc7609027bf3

#### Clint side filtering 

The user filter is simply a client-side filter because we anyway
currently have an in-memory table with client-side sorting and
pagination. There is no much practical value in server-side filter atm
The filter allows to filter-in by multiple users. It also allows to
filter items without creators

#### `created_by` filter isn't part of the query string

I introduced a separate state for this filter, similar to `tableSort`
and `pagination`. The reason why the `created_by` filter isn't a part of
the `searchQuery` is because we use `profile_id` which looks ugly for
the user, e.g. `u_mGBROF_q5bmFCATLXAcCwKa0k8JvONAwSruelyKA5E_0` so we
can't just add it to a user-facing search query like we do with tags.
`profile_id` is the correct identified we should use, `username` might
not be enough in some cases + `username` in cloud is unreadble account
number. The best alternative could be email, but it isn't always
available. This is why I decided to keep the `created_by` filter hidden
from the user behind the user picker popover and not expose it to the
search bar. The `created_by` filter is synced into the URL just like the
rest of the table state.

#### No distinguishing the current user

There is no logic in place that somehow highlights or renames the
current user in the filter. This can be added, but to limit the scope of
the MVP I didn't do it.


#### User filter component 

We use the `UsersProfilePopover` provided by the security team. We use
it without any overrides, so the component decides how to show users in
a popover. I noticed a small visual issue with an email truncation in a
smaller popover. Might be worth a separate issue.



### Testing 

1. Create multiple new users 
2. Login for each of them 
3. Created a dashboard for each of them 
4. Play with the filter 


#### Next Steps 


The next step would be to add a column with the user: 


![image
(5)](https://github.com/elastic/kibana/assets/7784120/9a3fe30e-6a41-4718-a7aa-53fbe0affd22)