Fixes flaky backfill tests (#181868)
Resolves https://github.com/elastic/kibana/issues/181778
Resolves https://github.com/elastic/kibana/issues/181861
Resolves https://github.com/elastic/kibana/issues/181862
Resolves https://github.com/elastic/kibana/issues/181875
Resolves https://github.com/elastic/kibana/issues/181840
Resolves https://github.com/elastic/kibana/issues/181893
Resolves https://github.com/elastic/kibana/issues/182029
Resolves https://github.com/elastic/kibana/issues/181907



## Summary

The schedule request tries to schedule backfills for multiple rule
types. The test was testing for a 403 Forbidden message containing a
specific rule type ID (the first one in the schedule request) but in
practice, the rule type that the authorization check runs on is
non-deterministic so could be any of the rule type IDs in the schedule
request. Changed the test to match the error pattern instead.

For the `GET` and `FIND` tests, we're getting 404s when we're expecting
200s so it seems like the backfill runs are completing before the get &
find request are issued. I increased the schedule for the backfill runs
so the jobs don't finish and get deleted before the tests run. I also
changed the pattern for matching `status` to allow for something other
than `pending` status (for backfill schedules that complete before the
GET request is made)

Flaky test runner runs: 
-
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/5804
-
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/5806

---------

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>