[ML] Fix retention of categorization example limits (#182103)
When updating a job's model memory limits via the UI, if
`categorization_examples_limit` has been set in `analysis_limits`, the
configured value is replaced by the default value of `4`.
This appears to be a
[bug](https://github.com/elastic/elasticsearch/issues/108068) in
elasticsearch and so this PR is a workaround to ensure the original
`categorization_examples_limit` is sent as part of the update to ensure
it isn't reset to the default.

Also fixes an issue in the advanced job wizard where setting
`categorization_examples_limit` in the JSON is not retained when saving
the JSON.

(cherry picked from commit 0842c608452d4b7f9a746f9a51395116e325272a)