[data.search] Fix unhandled promise rejections (#181785)
## Summary

Resolves https://github.com/elastic/kibana/issues/168957.

It turns out the underlying issue was resolved in
https://github.com/elastic/kibana/pull/170041 (unhandled errors when
deleting not being handled). However this still left it up to consumers
of `pollSearch` to be 100% sure they weren't leaking unhandled promise
rejections. This adds code directly to `pollSearch` that will handle
rejections if they aren't handled in the calling code. It also adds
tests to consumers of `pollSearch` to make sure they don't barf in the
case that the `cancel` function errors.

### Checklist

- [x] [Unit or functional
tests](https://www.elastic.co/guide/en/kibana/master/development-tests.html)
were updated or added to match the most common scenarios