[build/serverless] Remove validate cdn assets timeout
After uploading CDN assets we run a test to make sure all files exist.
There's currently a 1 second timeout on these requests, and this will
occasionally be exceeded.  This removes the timeout - testing the CDN for
latency is out of scope.  The timeout for this CI step is one hour, which
should be sufficient to avoid excessive runtime.