[Dashboard] Add "Creator" column to dashboard listing page (#182256)
## Summary

close https://github.com/elastic/kibana/issues/182253

This PR adds a creator column to the dashboard's listing page: 

![Screenshot 2024-05-02 at 09 29
12](https://github.com/elastic/kibana/assets/7784120/86ee27f9-0877-4574-95f8-11439f8a7d78)

![Screenshot 2024-05-02 at 09 29
17](https://github.com/elastic/kibana/assets/7784120/ec7836a9-b018-41c9-96b8-91004c5df78c)


- if none of the dashboards have a creator, the column is not rendered
- if creator is empty, then the cell is empty
- the column **is not sortable because** on the table level we only have
user ids and it doesn't make sense to sort by them. I prefer to not
block the dashboard rendering with the additional getAllUsers request to
make the column sortable. Now the users are only fetched on the cell
level during rendering, so we only fetch visible users. **I hope this is
an OK tradeoff that favors table's first paint and simplicity of
implementation**