Call bulkCreateSuppressedAlertsInMemory to ML rules
This adds all the parameters necessary to invoke this method (if
relevant) in the ML rule executor. Given the relative simplicity of the
ML rule type, I'm guessing that many of these values are
irrelevant/unused in this case, but I haven't yet investigated that.

Next step is to exercise this implementation against the FTR tests, and
see if the behavior is what we expect. Once that's done, we can try to
pare down what we need/use.

I also added some TODOs in the course of this work to check some
potential bugs I noticed.