Data telemetry: Widen tracked index patterns (#182563)
This PR adds a few tracked index patterns to the existing data
telemetry:
* `*logs*`
* `*logstash*`
* `*functionbeat*`
* `*heartbeat*`
* `*metricbeat*`
* `*filebeat*`

This is done for two reasons:
* As discussed
[here](https://github.com/elastic/observability-bi/issues/159#issuecomment-1914165610),
the `_meta` object in the mapping is not reliable to detect the shipper
* It's a common practice to prefix the regular names of indices (e.g.
`prod-filebeat`) - the existing patterns are not capturing these

## Considerations
* There is some overlap in this data (e.g. `*logs*` also matches
`*logstash*`). This is an accepted side effect and needs to be
considered when interpreting the results
* Changing the definitions of the logstash and *beat patterns would
cause a bump in the existing telemetry as its definition changes. To
avoid this, these patterns are added as separate entries (as
`generic-logstash` and so on).