[Fleet] Add `ignore_malformed: false` to `event.ingested` property in agent ID verification template (#182592)
## Summary

Set `ignore_malformed` to `false` on the `event.ingested` property that
we define as part of the Agent ID verification component template. Since
`event.ingested` is generated as part of the Fleet final pipeline, we
can expect that this value will never be malformed. This allows us to
enable synthetic source for Elastic Agent logs and save on storage.

Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>