[EDR Workflows][MKI] Skip processes cy test in MKI environment. (#182818)
In MKI we don't specify Fleet Server version, it's being provided to us
as part of the Serverless instance that these tests run against. It
caused issues with installing agents that are +2 versions ahead of the
FS. Therefore I went with an api call to
/api/fleet/agents/available_versions and install latest version of an
agent that is available. At this point it's 13.2 and it happens only if
MKI environment is detected

Since the `/components/agentbeat` took place in agent 14 it's not there
yet (in `13.2`). Hence the error we are seeing.