Implementing most of our FTR test functionality
I realized that most of these tests were using es_archiver to insert
anomalies into an index, but our tests were only ever using a single one
of those anomalies. In order to ensure these tests are independent of
the data in that archive, I've created and leveraged a helper to delete
all the persisted anomalies, and then use existing tooling to manually
insert the anomalies needed for our tests.

All of the current tests are green; there are just a few more
permutations that still need to be implemented.