es query rule - get time field from data view instead of rule (#182883)
## Summary

Previously it was possible to create a rule with a data view and change
the data view but the previous time field would still be referenced. Now
the time field is always pulled from the current data view.


Closes https://github.com/elastic/kibana/issues/182879

#### Release note

Fixed issue where an ES query rule could be created with a data view,
then the data view is changed but there's still a reference to the
previous data view's timestamp field. Now the timestamp field is always
taken from the currently configured data view.

---------

Co-authored-by: Davis McPhee <davis.mcphee@elastic.co>