Fixes Failing test: X-Pack Alerting API Integration Tests - Alerting - group4.x-pack/test/alerting_api_integration/spaces_only/tests/alerting/group4/alerts_as_data/alerts_as_data_flapping·ts - Alerting alerts_as_data alerts as data flapping should set flapping and flapping_history for flapping alerts over a period of time longer than the lookback (#182981)
Resolves https://github.com/elastic/kibana/issues/161477

## Summary

The test calls the `runSoon` API to run the rule a specific number of
times for the flapping tests. If the API returns a 204, the rule will be
run soon and this is what we expect. If the API returns a 200, likely
the rule is already running from a previous `runSoon` call. This adds a
retry to ensure the run soon call succeeds and kicks off a run.

Flaky test runner:
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/5908