Fix type errors in prebuilt rule cypress test
The addition of explicit alert suppresison fields on ML rules caused the
types in these tests to be incompatible.

Adding the `alert_suppression: undefined` (as the other rule types
do) was insufficient, as the omit() and type cast were doing extra
weirdness. Additionally, the fields that were being omitted are not
actual present (according to typescript) on the output of the
createRuleAssetSavedObject function (which has an implicit return type
as well).

I fixed this by preserving the type as if it hadn't been given a
Partial<> type via omit, and using a more explicit type (with `Omit`)
instead.