Fixing flaky ml test (#183139)
## Summary

Closes https://github.com/elastic/kibana/issues/177215

Changing ML rule functional test to stop the transform first, then
creating the rule. The rule runs when it is created and the test is
testing for `event.action: "open"` which is written on a new alert, so
we only want to rule to run once. Removed the unnecessary `runSoon`
call.

Flaky test runner:
https://buildkite.com/elastic/kibana-flaky-test-suite-runner/builds/5929