apm: use field API in service map aggregation
Improve robustness of the service map aggregation to missing field
mappings. Instead of failing when the mapping is missing, treat this
the same as the field value being missing in a document.

https://www.elastic.co/guide/en/elasticsearch/reference/current/script-fields-api.html:

> This API fundamentally changes how you access documents in Painless.
Previously, you had to access the doc map with the field name that you
wanted to access:
>
>   doc['my_field'].value
>
> Accessing document fields this way didn’t handle missing values or
missing mappings, which meant that to write robust Painless scripts, you
needed to include logic to check that both fields and values exist.