Add downstream dependency service name to logs and errors to improve alert insights (#183215)
Related: https://github.com/elastic/kibana/issues/183216

### Changes

- Exclude APM error docs from logs and retrieve APM errors separately
- Get sample `trace.id` from logs and apm errors, and retrieve the
downstream service name (if available)
- Minor prompt tweaks



### Scenario

When running the
[Otel-Demo](https://github.com/elastic/opentelemetry-demo) the
"checkout" service is killed on purpose. This causes the failure rate of
the frontend service to increase because is has a downstream dependency
on the checkout service. This in turn causes alerts to be triggered.

When the user navigates to the alerts details page, and opens the
insights they should be presented with the "checkout" service as the
root cause.

### Before

Before this change the alert insights did not capture that changes to
the `checkout` service was the root cause


![image](https://github.com/elastic/kibana/assets/209966/9235aec0-fb69-42fc-a692-7bd132ed819a)

### After


![image](https://github.com/elastic/kibana/assets/209966/a29c826a-708c-4be2-8fea-5e014a4fc41b)