[Connectors] [Jira] Fixed bug in additional fields  (#183285)
Fixes #183163

## Summary

In this scenario, the API was being called with `otherFields: ''`.

The backend expects `otherFields:
schema.nullable(schema.recordOf(...))`.

Initially, I thought about setting the value to `{}` when the user
cleared the form but we had that behavior in the past in another field
and I had to change it. 😅

Now I set the value to `null` and the API is called with valid values.

(cherry picked from commit 4ba10f02d37093966391958b1f77e708afa7372f)