[POC] Track time to first meaningful paint (#180309)
part of https://github.com/elastic/observability-dev/issues/3269 

#### Time to First Meaningful Paint (TTFMP)


Measures the time from the start of navigation to the point at which the
most meaningful element appears on the screen. TTFMP is instrumented
differently on each page as the most meaningful element varies from page
to page.

example: 
```
event_type : "performance_metric" and eventName: "kibana:plugin_render_time" 
```

#### Usage

To instrument TTFMP you need to `PerfomanceContextProvider` at the root
at the app and run `onPageReady` function once meaningful data are
fetched . The meaningful data can be one or more


```

import { usePerformanceContext } from '@kbn/ebt-tools';
const { onPageReady } = usePerformanceContext();

```

#### Current instrumentation

Based on telemetry, I instrumented the TTFMP for most viewed pages in
apm and infra:

1. /services 
2. /traces
3. /service-map
4. /hosts
5. infra /inventory

All pages except hosts have one component so the page ready once the
data is loaded for the component. For hosts I set it when the hosts list
table and the hosts number is loaded.
#### Telemetry  

Example dashboard of the metrics I sent from my local setup
-
[Dashboard](https://telemetry-v2-staging.elastic.dev/s/apm/app/dashboards#/view/f240fff6-fac9-491b-81d1-ac39006c5c94?_g=(filters:!(),refreshInterval:(pause:!t,value:60000),time:(from:now-7d,to:now)))


## Updates for reviewers
Previously the PR was checking both the component AND the time to first
meaningful paint. We decided to focus on the time to first meaningful
paint


### Notes
TTFMP is subject to change for each page based on what we're going to
define as meaningful

---------

Co-authored-by: kibanamachine <42973632+kibanamachine@users.noreply.github.com>