[CI] Avoid docker login if docker is not running (#183379)
## Summary
We've found that the command tries to log in to docker in cases when the
`docker` binary is available (installed) but not running on the devices.
This can happen, for example, on a device that's set up a bit
differently than our normal CI executors (e.g.: [bare metal
executors](https://buildkite.com/elastic/kibana-single-user-performance/builds/13383)).

This PR makes it sure that we only interact with docker in that step, if
it's running.

(cherry picked from commit 9d1a37d2db1c5616ad2619fde2d85afd0d102170)