Kbn 176803 cleanup 8.14 md5 check (#183416)
## Summary

Follow-up of #176803

Now that `8.14` has been branched and it is about to be released, we can
remove the test that ensures the `HASH_TO_VERSION_MAP` is aligned.

This PR also rolls back a few changes on the map that shouldn't have
been performed, as they belong to `8.15.0`.

(cherry picked from commit ed2e59fea95b39dcc853f948ce122544515370f0)