[MGMTXP] [Connectors] Added the `additional_info` field to the ITOM connector UI (#183380)
Fixes #171320

## Summary

The additional_info field for the ServiceNow ITOM connector already
existed in the backend. When creating an action it was being
pre-populated with some rule and alert data and sent to the backend with
the form.

In this PR I am making that field visible and editable by the user while
preserving the original default values.

<img width="1959" alt="Screenshot 2024-05-14 at 12 28 25"
src="https://github.com/elastic/kibana/assets/1533137/74f81ff0-a141-4744-b86c-3a7146796342">

## Release notes

Added support for the additional info field in the ServiceNow ITOM
connector.