[Security Solution] Using Chrome on CI for Cypress (#183515)
## Summary

Fixes: https://github.com/elastic/kibana/issues/183098

Lately we have experienced some issues with our Cypress tests, where we
are seeing that tests are passing but the output is marked as a failure.

@patrykkopycinski has done a great job trying to fix the issue but
lately we have seen that is still reproducible.

The issue seems to be related directly with electron, so in this PR we
reenabling chrome on CI to try to fix it.

In order to do so, some adjustments has been introduced to woraround the
chrome out of memory crashes that we faced before and was the main
reason to move to electron.

---------

Co-authored-by: Patryk Kopycinski <contact@patrykkopycinski.com>
(cherry picked from commit a72ddc095903bcdc6441e17706be627c661da870)