Fix nondeterministic ordering of alerts
Our assertions assume a specific ordering for the alerts generated by
the rule, but the timestamps are identical on the underlying documents
so it's a guess as to which one comes back "first" in these tests. This
flakiness was caught by the flaky test runner.

To ensure the ordering here, we instead sort our alerts by
`kibana.alert.suppression.docs_count`, which will be 0 and 1 for our two
alerts, respectively.