[Entity Analytics] Move scripted metric painless scripts to static file & remove category based weighting (#182038)
## Summary

The scripted metric aggregation is being deprecated but an exception is
being made for our team, moving the painless to static files allows for
this exception to be made. We have had to remove category weighting to
make the script less dynamic, weights weren't used anyway so not a
breaking change.

The scripts are loaded once when they are first used and then cached. A
unit test verifies the content of the script hasnt changed.

Here is a diff of the scripted metric before and after
https://www.diffchecker.com/gefuBoYK/

---------

Co-authored-by: Mark Hopkin <mark.hopkin@elastic.co>
Co-authored-by: Kibana Machine <42973632+kibanamachine@users.noreply.github.com>